xdevops-prompt-optimizer/requirements.txt
openai
python-dotenv
loguru
pydantic
tiktoken